+++ device-minion-runtime/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- [SW-16238] Adding support for kernel environment variables.
 ### Changed
 - Updating the gitlab CI repo pointer to latest.
 - [SW-15697] Replacing threshold_data with threshold value in trace buffer full notification.